The last commit comment did not do justice.  Here is the comment that should have been attached to it.
ue4 does not recognize enumerations. Enumerations are still good to use, but to avoid compiler warnings and errors, we must cast all enumerations as (int)

The binding only needed to be performed once. I elevated the binding functions to SetupPlayerInputComponent.  Once I did this, the SetupControlsSwim and SetupControlsPropel were identical.  I renamed those functions to MapMotionControllersToHands.

Restored the keyboard mouse capability of looking up and down.

Somehow OPTIMIZATION was turned on! Optimization is rarely a good thing. The highest level was set and was removing functions and variables that the compiler "believed" to not be used.  This took hours to debug :(

Re-integrated the software architecture and pickup class implemention that was completed last week. These were somehow not merged in.

NOTE:
1) This check breaks the inventory which is a known and fairly straight forward issue to resolve.
2) I was unable to recover the same functionality I had before with the pickup. Something is still wrong.