I have been researching ue4 animation and have found little support for pure c++ animation.  In fact, it appears the only way to “properly” animate ue4 objects is to use blueprints and event graphs.  However, many of our animation simple. Such as open a door or a tool box.
Although, there is lack of documentation for pure c++ implementation of animations, all GUI interfaces usually compile down to straight forward c implementation. Meaning, if you can do something in the blueprint event graphs, you should be able to do it in code.  So I watched a couple animation tutorials (in blueprints) and I think it will pretty straight forward to animate the items we need.

The first discovery is that every animation they use the EventTick. Which is the Tick function in code.  However, we need to be careful in that we do not want to call Tick for every frame on every object. Nevertheless, we do need to call Tick for an object during the animation process.

Therefore, for each Interactable object, I have added the Tick() function to the Interactable class and disabled the call to it by default.  If an interactable requires animation, the derived class can override the virtual ExecuteAnimation() for custom animations.  Then upon action (like a trigger pull for example) the derived class would EnableAnimation(), then animate the object in the overridden ExecuteAnimation() function.  (See Toolbox as an example).